Relocate VirtualMachineOrGroup interface and fix VM/VMG get functions receiver
